converter, storage: extract volume conversion logic
Move all volume converter functions (Convert_v1_Volume_To_api_Disk,
Convert_v1_Hotplug_Volume_To_api_Disk, Convert_v1_Missing_Volume_To_api_Disk,
SetDiskDriver, CBT helpers, PVC/DV/HostDisk/CloudInit/Sysprep/EmptyDisk/
ContainerDisk/Ephemeral/DownwardMetric converters) from converter.go
to storage/volumes.go
