+++ pkg/virt-launcher/virtwrap/converter/BUILD.bazel
@@ -10,12 +10,7 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/cloud-init:go_default_library",
-        "//pkg/config:go_default_library",
-        "//pkg/container-disk:go_default_library",
-        "//pkg/emptydisk:go_default_library",
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
-        "//pkg/host-disk:go_default_library",
         "//pkg/network/vmispec:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/hardware:go_default_library",
