+++ pkg/virt-launcher/virtwrap/converter/storage/BUILD.bazel
@@ -11,10 +11,16 @@ go_library(
         "generated_mock_directio.go",
         "paths.go",
         "virtiofs.go",
+        "volumes.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/storage",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/cloud-init:go_default_library",
+        "//pkg/config:go_default_library",
+        "//pkg/container-disk:go_default_library",
+        "//pkg/emptydisk:go_default_library",
+        "//pkg/host-disk:go_default_library",
         "//pkg/os/disk:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/safepath:go_default_library",
@@ -45,6 +51,7 @@ go_test(
         "driver_test.go",
         "storage_suite_test.go",
         "virtiofs_test.go",
+        "volumes_test.go",
     ],
     race = "on",
     deps = [
